Separate methods from MockUtils into CategorySyncMockUtils.
